+++ CHANGELOG.md
@@ -4,6 +4,10 @@
 
 ### Improvements
 - Better cache handling with Etags ([#1097](../../pull/1097))
+- Reduce duplicate computation of slow cached values ([#1100](../../pull/1100))
+
+### Bug Fixes
+- Tile serving can bypass loading a source if it is in memory ([#1102](../../pull/1102))
 
 ### Changes
 - On the large image item page, guard against overflow ([#1096](../../pull/1096))